feat: Refeactor into more generic components
